Connect the login and sign in page with the database
Try to connect he login and sign in page with the database but it still have some bugs related to the firebase. Not sure it is the code error or the connection between the project and the firebase. I am going to resolve it tomorrow. Hope it works tomorrow.

Add a toast message for successfully register or not.